+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Cluster discovery dynamically configured with `apply`.
 - Discovery settings apply and delete events.
+- Introduce `core.replicante.io/debug` actions.
 - List and delete `DiscoverySettings` objects (API and `replictl`).
 - Orchestrate reports for details on the latest attempt to sync a cluster.
 - Orchestrator actions, executing during cluster orchestration.